chore(ci): do not use personal token in flow
Github workflows that use the provided GITHUB_TOKEN cannot trigger other
workflows. However, workflows that run on `workflow_dispatch` are an
exception to that. Therefore, we can stop using our own Personal Token
here and resort to using the one provided by Github, which is more
secure and easier to maintain.

See:
https://docs.github.com/en/actions/writing-workflows/choosing-when-your-workflow-runs/triggering-a-workflow#triggering-a-workflow-from-a-workflow