refactor(setlist): move min setlist songs to repo
By doing so, we keep the deserializer object simple having a single
responsibility. It is up to the repository, now, to only retrieve the
setlists with a minimum amount of songs.

Also took the chance to make the setlistfm response class private, as it
is good practise to only expose the ones that are actually used outside
the module.